Week_3/Day_1/Hackathon/pgsqlCopyEditbook_review_api/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib
pydantic
python-dotenv

